third_party/WebKit/LayoutTests/fast/dom/shadow/part-pseudo-element-css-text-expected.txt

Test for cssText of ':host()' rule.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.getElementById('style1').sheet.cssRules.item(0).cssText is "div > .foo::part(target) { display: block; }"
PASS document.getElementById('style-invalid').sheet.cssRules.length is 0
PASS successfullyParsed is true

TEST COMPLETE
